Create API package. Move API files to package.
